+++ src/include/drivers/video_term.h
@@ -17,21 +17,21 @@ struct vterm;
 
 struct vterm_ops {
 	void (*init)(struct vterm *t);
-	void (*cursor)(struct vterm *t, uint32_t x, uint32_t y);
-	void (*putc)(struct vterm *t, char ch, uint32_t x, uint32_t y);
-	void (*clear)(struct vterm *t, uint32_t x, uint32_t y,
-			uint32_t width, uint32_t height);
-	void (*move)(struct vterm *t, uint32_t sx, uint32_t sy, uint32_t width,
-			uint32_t height, uint32_t dx, uint32_t dy);
-	void (*scroll)(struct vterm *t, int32_t delta);
+	void (*cursor)(struct vterm *t, unsigned short x, unsigned short y);
+	void (*putc)(struct vterm *t, char ch, unsigned short x, unsigned short y);
+	void (*clear)(struct vterm *t, unsigned short x, unsigned short y,
+			unsigned short width, unsigned short height);
+	void (*move)(struct vterm *t, unsigned short sx, unsigned short sy, unsigned short width,
+			unsigned short height, unsigned short dx, unsigned short dy);
+	void (*scroll)(struct vterm *t, short delta);
 };
 
 struct vterm {
-	uint32_t cur_x;
-	uint32_t cur_y;
-	uint32_t back_cx, back_cy;
-	uint32_t width;
-	uint32_t height;
+	unsigned short cur_x;
+	unsigned short cur_y;
+	unsigned short back_cx, back_cy;
+	unsigned short width;
+	unsigned short height;
 
 	const struct vterm_ops *ops;
 	void *data;
@@ -40,7 +40,7 @@ struct vterm {
 	struct vtesc_executor executor;
 };
 
-extern void vterm_init(struct vterm *t, uint32_t width, uint32_t height,
+extern void vterm_init(struct vterm *t, unsigned short width, unsigned short height,
 		const struct vterm_ops *ops, void *data);
 extern void vterm_putc(struct vterm *t, char ch);
